Renamed directory.
