+++ src/app/browse-by/browse-by-page.module.ts
@@ -4,7 +4,7 @@ import { BrowseByModule } from './browse-by.module';
 import { ItemDataService } from '../core/data/item-data.service';
 import { BrowseService } from '../core/browse/browse.service';
 import { BrowseByGuard } from './browse-by-guard';
-import { SharedBrowseByModule } from '../shared/shared-browse-by/shared-browse-by.module';
+import { SharedBrowseByModule } from '../shared/browse-by/shared-browse-by.module';
 
 @NgModule({
   imports: [
