+++ src/app/shared/browse-by/browse-by.component.scss
