+++ src/themes/custom/eager-theme.module.ts
@@ -45,7 +45,7 @@ import { CommunityListElementComponent } from './app/shared/object-list/communit
 import { CollectionListElementComponent} from './app/shared/object-list/collection-list-element/collection-list-element.component';
 import { ResultsBackButtonComponent } from './app/shared/results-back-button/results-back-button.component';
 import { CollectionDropdownComponent } from './app/shared/collection-dropdown/collection-dropdown.component';
-import { SharedBrowseByModule } from '../../app/shared/shared-browse-by/shared-browse-by.module';
+import { SharedBrowseByModule } from '../../app/shared/browse-by/shared-browse-by.module';
 import { ResultsBackButtonModule } from '../../app/shared/results-back-button/results-back-button.module';
