+++ src/themes/custom/lazy-theme.module.ts
@@ -117,7 +117,7 @@ import { BrowseByTitlePageComponent } from './app/browse-by/browse-by-title-page
 import {
   ExternalSourceEntryImportModalComponent
 } from './app/shared/form/builder/ds-dynamic-form-ui/relation-lookup-modal/external-source-tab/external-source-entry-import-modal/external-source-entry-import-modal.component';
-import { SharedBrowseByModule } from '../../app/shared/shared-browse-by/shared-browse-by.module';
+import { SharedBrowseByModule } from '../../app/shared/browse-by/shared-browse-by.module';
 import { ResultsBackButtonModule } from '../../app/shared/results-back-button/results-back-button.module';
 
 const DECLARATIONS = [
