+++ src/themes/dspace/lazy-theme.module.ts
@@ -57,7 +57,7 @@ import { ComcolModule } from '../../app/shared/comcol/comcol.module';
 import { RootModule } from '../../app/root.module';
 import { BrowseByPageModule } from '../../app/browse-by/browse-by-page.module';
 import { ResultsBackButtonModule } from '../../app/shared/results-back-button/results-back-button.module';
-import { SharedBrowseByModule } from '../../app/shared/shared-browse-by/shared-browse-by.module';
+import { SharedBrowseByModule } from '../../app/shared/browse-by/shared-browse-by.module';
 
 const DECLARATIONS = [
 ];
